Add abstraction on Services
