+++ src/main/java/com/edu/cs554/airlinesreservationsystem/service/Implementation/LoginServiceImpl.java
@@ -1,0 +1,50 @@
+package com.edu.cs554.airlinesreservationsystem.service.Implementation;
+
+import com.edu.cs554.airlinesreservationsystem.dto.AuthRequest;
+import com.edu.cs554.airlinesreservationsystem.model.User;
+import com.edu.cs554.airlinesreservationsystem.repository.UserRepository;
+import com.edu.cs554.airlinesreservationsystem.service.LoginService;
+import com.edu.cs554.airlinesreservationsystem.util.JwtUtil;
+import lombok.AllArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.stereotype.Service;
+
+@Service
+@AllArgsConstructor
+public class LoginServiceImpl implements LoginService {
+
+    @Autowired
+    private JwtUtil jwtUtil;
+
+    @Autowired
+    private AuthenticationManager authenticationManager;
+
+    @Autowired
+    private UserRepository userRepository;
+
+    public String authenticate(AuthRequest request){
+        authenticationManager.authenticate(new UsernamePasswordAuthenticationToken(request.getUserName(), request.getPassword()));
+        return jwtUtil.generateToken(request.getUserName());
+    }
+
+    public String getLoggedInUser(){
+        String username;
+        Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (principal instanceof UserDetails) {
+            username = ((UserDetails)principal).getUsername();
+        } else {
+            username = principal.toString();
+        }
+
+        return username;
+    }
+
+    public User loggedInUser() {
+        User user = userRepository.findByUsername(getLoggedInUser());
+        return user;
+    }
+}
