+++ src/main/java/com/edu/cs554/airlinesreservationsystem/service/Implementation/UserServiceImpl.java
@@ -1,0 +1,42 @@
+package com.edu.cs554.airlinesreservationsystem.service.Implementation;
+
+import com.edu.cs554.airlinesreservationsystem.model.User;
+import com.edu.cs554.airlinesreservationsystem.repository.UserRepository;
+import com.edu.cs554.airlinesreservationsystem.service.UserService;
+import lombok.AllArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.ArrayList;
+
+@Service
+@Transactional
+@AllArgsConstructor
+public class UserServiceImpl implements UserService {
+
+    @Autowired
+    private final BCryptPasswordEncoder bCryptPasswordEncoder;
+
+    @Autowired
+    private UserRepository userRepository;
+
+    @Override
+    public UserDetails loadUserByUsername(String userName) throws UsernameNotFoundException {
+        User user = userRepository.findByUsername(userName);
+        return new org.springframework.security.core.userdetails.User(user.getUsername(), user.getPassword(), new ArrayList<>());
+    }
+
+    public User encodePassword(User user){
+
+        String encodedPassword = bCryptPasswordEncoder.encode(user.getPassword());
+        user.setPassword(encodedPassword);
+
+        return user;
+    }
+
+}
